Refactored CreateTranslationDrafts bkg task
